+++ tests/test-legacy-vue2/vite.config.js
@@ -1,0 +1,10 @@
+/*!
+ * SPDX-License-Identifier: CC0-1.0
+ * SPDX-FileCopyrightText: 2026 Nextcloud GmbH and Nextcloud contributors
+ */
+
+import { createAppConfig } from '@nextcloud/vite-config'
+
+export default createAppConfig({
+	app: 'src/main.js',
+})